# Speed up feature_management_spec.rb by removing use of browser