```
bundle exec rspec spec/lib/feature_management_spec.rb
# BEFORE
Finished in 2 minutes 38.5 seconds (files took 3.96 seconds to load)
# AFTER
Finished in 0.76372 seconds (files took 7.82 seconds to load)
```
about a 200x speedup