@zachmargolis 2022-09-22
> Nice find! Was there a particular way you discovered this, or just stumbled across it? I'm curious now if there are others.

I sorted by the keys in knapsack report, and looked for things outside of `spec/features` and this one was a very big surprise